feat(ci): :sparkles: Add semantic-release automation for versioning
This commit introduces semantic-release automation to handle semantic versioning from conventional commits. It sets up a GitHub Actions workflow to build the Rust project and run semantic-release. Additionally, it configures semantic-release with necessary plugins and scripts to update the version in Cargo.toml, generate a changelog, and commit these changes back to the repository.

- Added release.yml for GitHub Actions workflow
- Configured semantic-release in package.json and release.config.js
- Added update_version.sh script to update version in Cargo.toml.
- Added `ci` as a scope in workspace settings

This setup will streamline the release process, ensuring consistent versioning and changelog updates with every merge to the main branch.

📦 target/release/zk_whitelist